pyproject.toml
[project]
name = "pisensor"
version = "0.1.0"
description = ""
authors = [
    {name = "好耶！",email = "86039582+cloudw233@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "orjson (>=3.10.16,<4.0.0)",
    "smbus (>=1.1.post2,<2.0)",
    "gpiod (>=2.3.0,<3.0.0)",
    "pyserial (>=3.5,<4.0)",
    "pynmea2 (>=1.19.0,<2.0.0)",
    "pigpio (>=1.78,<2.0)",
    "adafruit-circuitpython-dht (>=4.0.7,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "gpiozero (>=2.0.1,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "attrs (>=25.3.0,<26.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "smbus2 (>=0.5.0,<0.6.0)",
    "rpi-gpio (>=0.7.1,<0.8.0)",
    "lgpio (>=0.2.2.0,<0.3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "spidev (>=3.7,<4.0); sys_platform == \"linux\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[tool.poetry]
package-mode = false
